Fix resolutiondatetime way off in the future
Fix the handling of creationdate and resolutiondate. We were sending things to Date to long and back to Date. Fix that by handling things internally in Action as the structured data, and then reformatting it the way we want. This requires some null handling (via `Optional<>`) which is honestly better separation of concerns anyway.

Also better handles resolution date. As discovered in #58, this isn't a real field. It's a shortcut for the date the ticket was resolved.