Fix: Datetimes way off in the future
I think the main problem was we were sending things to Date to long and
back to Date. Fix that by handling things internally in Action as the
structured data, and then reformatting it the way we want. This requires
some null handling (via `Optional<>`) which is honestly better
separation of concerns anyway.