CustomField Parser can remove empty strings
The CustomFieldDefinition.SplitRule is pretty simple; it just replaces
every single instance of the matched pattern with the replacement. The
problem is that data that unexpectedly does not match at the beginning
or the end of the input can produce output with a leading or trailing
separator. If you're splitting on NON_NUMBER, an input like "abc123def"
produces an output like "|123|". If the field is strictly typed as an
integer, Imhotep will reject it. Thus we need to be configured to
automatically remove those.

Fixes #63