Correctly decodes projectkey from issuekey
We were previously getting the projectkey from the Project section in
the fields. This worked great until we wanted to see the project key as
it was in history. We were always getting the current one.

This changes to always decoding the project key from the leading
characters in the issuekey.